requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0
datasets>=2.14.0
scikit-learn>=1.3.0
pydantic>=2.0.0
